pyproject.toml
[project]
name = "cashevide-api"
version = "0.1.0"
description = "An open-source initiative to ensure payment security and client transparency for freelancers. Built with transparency, powered by Python & Django."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "boto3>=1.42.61",
  "dj-database-url>=3.1.2",
  "django>=6.0.1",
  "django-cors-headers>=4.9.0",
  "django-storages>=1.14.6",
  "djangorestframework>=3.16.1",
  "djangorestframework-simplejwt>=5.5.1",
  "drf-spectacular>=0.29.0",
  "gunicorn>=25.1.0",
  "phonenumbers>=9.0.23",
  "pillow>=12.1.1",
  "psycopg2-binary>=2.9.11",
  "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
  "django-stubs>=5.2.9",
  "django-stubs-ext>=5.2.9",
  "django-types>=0.22.0",
  "djangorestframework-stubs>=3.16.7",
]

[tool.pyright]
reportIncompatibleVariableOverride = false
